requirements.txt
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.4
discord.py==2.2.2
python-dotenv==0.21.1
pymongo==4.4.0.dev1
tiktoken==0.3.1